# apps/api/internal/spec/openapi.yaml
openapi: 3.0.3
info:
  title: UFO Hub API
  version: 0.5.0
  license:
    name: BSD-3-Clause
    url: https://opensource.org/license/bsd-3-clause
  description: >
    Public beta Hub HTTP API for UFO (Unified Fleet Orchestrator). Endpoints and
    payloads may still change before 1.0; match clients to tagged releases and
    their release notes.
    Three surfaces: auth (public), the tenant/UI surface (httpOnly access JWT cookie
    or bearer token), and the rover surface (per-rover connection token). Source of truth for the
    protocol; maintained by hand alongside the code (generated clients are a follow-up).
servers:
  - url: /
security: []
...
paths:
...
  /v1/rovers/{id}:
    get:
      tags: [rover]
      operationId: getRover
      summary: Read a rover (fleet session, or the rover's own connection token)
      security: [{ accessCookie: [] }, { userBearer: [] }, { roverConnectionToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/RoverVersion" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rover" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "426": { $ref: "#/components/responses/UpgradeRequired" }
    patch:
      tags: [ui]
      operationId: patchRover
      summary: Update a rover's fields
      description: >
        Owner/admin may set name, tags, and units; the rover's own connection
        token may set auto_tags. Fields outside the caller's scope are rejected.
      security: [{ accessCookie: [] }, { userBearer: [] }, { roverConnectionToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/RoverVersion" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1 }
                units: { type: integer, minimum: 1, maximum: 100 }
                auto_tags: { type: array, items: { type: string } }
                tags: { type: array, items: { type: string } }
      responses:
        "204": { description: Updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "426": { $ref: "#/components/responses/UpgradeRequired" }
    delete:
      tags: [ui]
      operationId: deleteRover
      summary: Revoke a rover (owner/admin, or the rover's own connection token; its connection token stops working)
      security: [{ accessCookie: [] }, { userBearer: [] }, { roverConnectionToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/RoverVersion" }
      responses:
        "204": { description: Revoked }
        "403": { $ref: "#/components/responses/Forbidden" }
        "426": { $ref: "#/components/responses/UpgradeRequired" }
  /v1/rovers/{id}/stream:
    get:
      tags: [rover]
      operationId: roverStream
      summary: Server-Sent Events stream of config updates and revocation for the rover
      description: >
        SSE stream (the rover's own connection token). The server pushes
        `config` events `{"name":"...","units":N,"tags":[]}` whenever the
        rover's settings change, and a `revoke` event when the rover is revoked
        (after which the connection token stops working). Auto-reconnect is the
        rover's responsibility.
      security: [{ roverConnectionToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/RoverVersion" }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema: { type: string }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "426": { $ref: "#/components/responses/UpgradeRequired" }
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: ufo_access
      description: "Browser auth: httpOnly access JWT cookie."
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access JWT for non-browser clients.
    enrollmentCode:
      type: http
      scheme: bearer
      description: Enrollment code from the Rovers panel; exchanged once for a connection token.
    roverConnectionToken:
      type: http
      scheme: bearer
      description: Per-rover connection token returned by enrollment.
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Public resource id.
      schema: { type: string, format: uuid }
...
    RoverVersion:
      name: X-UFO-Rover-Version
      in: header
      required: false
      description: Required when authenticating with an enrollment code or rover connection token; ignored for user-authenticated requests.
      schema:
        type: string
        example: "0.5.0"
  responses:
    BadRequest:
      description: Bad request
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Not a member of the fleet
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    UpgradeRequired:
      description: Rover version is outside this Hub's supported range.
      headers:
        X-UFO-Rover-Min-Version:
          description: Minimum UFO rover CLI version accepted by this Hub, when configured.
          schema: { type: string }
        X-UFO-Rover-Max-Version:
          description: Maximum UFO rover CLI version accepted by this Hub, when configured.
          schema: { type: string }
        X-UFO-Hub-Version:
          description: Hub version that rejected the rover request.
          schema: { type: string }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
...
    Rover:
      type: object
      properties:
        id: { type: string, format: uuid }
        fleet_id: { type: string, format: uuid }
        fleet_name: { type: string }
        name: { type: string }
        status: { type: string, enum: [online, full, offline], description: full means all units are occupied }
        units: { type: integer, minimum: 1, maximum: 100, description: concurrent operations this rover can process }
        running_units: { type: integer, minimum: 0, description: active operations currently accepted by this rover }
        auto_tags: { type: array, items: { type: string }, description: "rover-detected (pilot:*, os:*, arch:*)" }
        tags: { type: array, items: { type: string }, description: user-set dispatch tags }
        metadata: { type: object, additionalProperties: true, readOnly: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
...